storage_proxy: config: add external_speculative_retry_timeout_in_ms
Before this change, if a read executor had just enough targets to
achieve query's CL and there was a connection drop (e.g. node failure),
the read executor waited for the entire request timeout to give drivers
time to execute a speculative read in a meantime. Such behavior don't
work well when a very long query timeout (e.g. 1800s) is set, because
the unfinished request blocks topology changes.

This change implements a new mechanism to prevent from topology
blocking after a connection drop, but also gives drivers enough time
to execute a speculative read. If a query is known to be failed due to
disconnections, read executor instead of waiting for the query timeout,
will wait for the external_speculative_retry_timeout and then fail the
request with read_failure_exception.

This change:
 - Add external_speculative_retry_timeout_in_ms config parameter with
   default value 20s
 - Implement use of external_speculative_retry_timeout_in_ms in
   digest_read_resolver
 - Implement test_external_speculative_retry to verify the new code

Closes scylladb/scylladb#21831