mapreduce: add tablet-aware dispatching algorithm
The main motivation for this change is reducing the time when ERM is
being kept by mapreduce service, to make sure that long aggregate
queries don't block topology operations. As ScyllaDB is generally moving
towards tablets, and using tablets makes work dispatching easier, the
decision was made to prepare the new algorithm only for tablets.

The new algorithm can be summarized as follows:
 1. Prepare `tablet_replica` -> `partition_range` mapping where values
    cover the entire space.
 2. In parallel, for each `tablet_replica`, take two partition ranges
    and dispatch it to the node hosting the replica. ERM is released and
    re-acquired in each iteration, so the destination (i.e.
    tablet_replica) can change for each partition range, and in such
    case the partition range is assigned to the proper tablet_replica.

In step 1, the main difference in comparison to the old algorithm
(dispatch_to_vnodes) is that partition ranges are assigned to
tablet_replica that is <host,shard> pair, instead of just the host.

In step 2, the main difference is that the work is divided into smaller
batches before dispatching, and ERM is released and re-acquired for each
batch.

In the current implementation, each node can correctly handle every
partition range, even if mapreduce supercoordinator keeps no ERM, and
the range it is absent locally. That is because
`mapreduce_service::execute_on_this_shard` creates a new pager that
coordinates the partition range read, including obtaining its own ERM.
However, every partition range that is absent locally is handled by
shard 0, therefore proper routing of partition ranges is required to
avoid shard 0 overload. That is why, in step 2, ERM is being kept during
each batch processing, and the tablet_replica is refreshed for each
processed range.

The tablet split scenario is not well handled in the
implementation introduced in this commit - after a split, the entire
pre-split partition range is sent to the node that hosts tablet_replica
containing end_token of the range. In result, a considerable part of the
range is handled by shard 0. This problem is solved later in this patch
series, by introducing `shard_id` in mapreduce_request.

Finally, the algorithm uses `parallel_for_each` for each tablet_replica.
If new replicas appear (e.g. after node addition), the parallel
iteration is stopped, and resumed on new set of tablet_replicas. It is
possible to avoid stopping and resuming but it would further complicate
the code, and the performance gain would be rather marginal.

Ref. scylladb#21831