mapreduce: add shard_id to mapreduce request
If a partition range is absent locally,
partition_ranges_owned_by_this_shard assigns it to shard 0.

This commit introduces shard_id in mapreduce_service, which overrides
this behavior. When shard_id is set, the entire partition range
is handled by the specified shard. The new tablet-aware mapreduce
algorithm, introduced in dispatch_to_tablets, balances the workload
across shards. The changes introduced in this commit ensure that
the balance is preserved, even during events such as tablet splits.

Fixes: scylladb#21831